refactor + delete old screen
